backend/scraper/requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
selenium==3.141.0
urllib3==1.24.1
elasticsearch-dsl>=6.0.0,<7.0.0